More buttons and activities (#13)
* More buttons and activities

* Hardware problems

The sensor callback interval is not set properly. So use 10ms for
the sampling rate.

I realized that the sensor callbacks are not properly invoked
while the screen is off. So I added some privileges to the manifest
file and `efl_util_set_window_screen_mode(EFI_UTIL_SCREEN_MODE_
DEFAULT)` to `btn_clicked_cb`.


The timestamp now prints relative timing instead of absolute time,
which is good for debugging.
* File format